multi now includes decimal

+++ src/main/java/com/insurgencedev/shopguiplusaddon/listeners/ShopGUIPlusEventListener.java
@@ -53,7 +53,7 @@ public final class ShopGUIPlusEventListener implements Listener {
         return false;
     }
 
-    private long calculateAmount(double amount, double multi) {
-        return (long) (amount * (multi < 1 ? 1 + multi : multi));
+    private double calculateAmount(double amount, double multi) {
+        return amount * (multi < 1 ? 1 + multi : multi);
     }
 }
